Extract DPM wave API module
